argv working

+++ tests/correct3/test03-while.py
@@ -1,4 +1,4 @@
-x::int = 4
+x = 4
 while x:
     print x
     x = x - 1
